simple ui and logic finishes version 2
