+++ src/components/HorizontalTimeline.tsx
@@ -1,197 +1,262 @@
-// import React, { useState } from "react";
+import React, { useState } from "react";
+import type { HorizontalTimelineProps, MilestoneProgress } from "../Interfaces/MilestoneModel";
+import { updateMilestoneStatus } from "../services/progressService";
+import formatDate from "../utils/DateFormater";
 
-// interface HorizontalTimelineProps {
-//   totalItems?: number;
-// }
+export default function HorizontalTimeline({ 
+  milestones = [],
+  userId
+}: HorizontalTimelineProps): React.ReactElement {
+  const [expandedSquare, setExpandedSquare] = useState<number | null>(null);
+  const [localMilestones, setLocalMilestones] = useState<MilestoneProgress[]>(milestones);
 
-// export default function HorizontalTimeline({ 
-//   totalItems = 10
-// }: HorizontalTimelineProps): React.ReactElement {
-//   const items = Array.from({ length: totalItems }, (_, index) => index + 1);
-//   const [clickedItems, setClickedItems] = useState<Set<number>>(new Set());
-//   const [expandedSquare, setExpandedSquare] = useState<number | null>(null);
+  // Update local state when milestones prop changes
+  React.useEffect(() => {
+    setLocalMilestones(milestones);
+  }, [milestones]);
 
-//   const timelineTexts = [
-//     "Induction", "Workshop 1", "Workshop 2", "Workshop 3", "Workshop 4", 
-//     "Workshop 5", "Workshop 6", "Workshop 7", "Portfolio", "Final Interview"
-//   ];
+  // Add debugging and ensure milestones is an array
+  console.log("Milestones received:", milestones);
+  console.log("Milestones type:", typeof milestones);
+  console.log("Is array:", Array.isArray(milestones));
 
-//   const handleSquareClick = (item: number) => {
-//     setExpandedSquare(expandedSquare === item ? null : item);
-//   };
+  // Ensure milestones is always an array
+  const milestonesArray = Array.isArray(localMilestones) ? localMilestones : [];
 
-//   const handleMarkAsDone = (item: number) => {
-//     setClickedItems(prev => {
-//       const newSet = new Set(prev);
-//       if (newSet.has(item)) {
-//         newSet.delete(item);
-//       } else {
-//         newSet.add(item);
-//       }
-//       return newSet;
-//     });
-//     setExpandedSquare(null);
-//   };
+  // Convert milestones to a format that includes completion status
+  const milestoneItems = milestonesArray.map((milestone, index) => ({
+    id: milestone.Milestone.id,
+    title: milestone.Milestone.title,
+    description: milestone.Milestone.description,
+    isCompleted: milestone.Status === "completed",
+    completionDate: milestone.CompletedAt,
+    index: index + 1
+  }));
+
+  const handleSquareClick = (milestoneId: number) => {
+    setExpandedSquare(expandedSquare === milestoneId ? null : milestoneId);
+  };
+
+  const handleMarkAsDone = async (milestoneId: number) => {
+    if (!userId) {
+      console.error("User ID is required to update milestone status");
+      return;
+    }
+
+    const currentMilestone = localMilestones.find(m => m.Milestone.id === milestoneId);
+    if (!currentMilestone) return;
+
+    const newStatus = currentMilestone.Status === "completed" ? "in-progress" : "completed";
+
+    // Update local state immediately for instant UI feedback
+    setLocalMilestones(prevMilestones => 
+      prevMilestones.map(milestone => {
+        if (milestone.Milestone.id === milestoneId) {
+          return {
+            ...milestone,
+            Status: newStatus,
+            CompletedAt: newStatus === "completed" ? new Date().toISOString() : undefined
+          };
+        }
+        return milestone;
+      })
+    );
+    
+    try {
+      // Make API call to persist the change
+      await updateMilestoneStatus(userId, milestoneId, newStatus);
+      console.log(`Successfully updated milestone ${milestoneId} to ${newStatus}`);
+    } catch (error) {
+      console.error("Failed to update milestone status:", error);
+      
+      // Revert local state if API call fails
+      setLocalMilestones(prevMilestones => 
+        prevMilestones.map(milestone => {
+          if (milestone.Milestone.id === milestoneId) {
+            return {
+              ...milestone,
+              Status: currentMilestone.Status,
+              CompletedAt: currentMilestone.CompletedAt
+            };
+          }
+          return milestone;
+        })
+      );
+      
+      alert("Failed to update milestone status. Please try again.");
+    }
+    
+    setExpandedSquare(null);
+  };
 
 
-//   return (
-//     <div className="w-full max-w-7xl mx-auto p-8">
-//       <div className="flex flex-wrap items-start justify-between" style={{ paddingTop: '40px', paddingBottom: '120px' }}>
-//         {items.map((item, index) => {
-//           const isTopPosition = index % 2 === 0; // Alternate top/bottom
+  return (
+    <div className="w-full max-w-7xl mx-auto p-6">
+      {milestoneItems.length === 0 ? (
+        <div className="text-center py-8">
+          <p className="text-gray-600">No milestones available yet.</p>
+        </div>
+      ) : (
+        <div className="flex flex-wrap items-start justify-between" style={{ paddingTop: '40px', paddingBottom: '120px' }}>
+          {milestoneItems.map((milestone, index) => {
+          const isTopPosition = index % 2 === 0; // Alternate top/bottom
           
-//           return (
-//             <div key={item} className="relative flex-shrink-0 mb-8" style={{ width: '100px', height: '280px', minWidth: '80px' }}>
-//               {/* Square on top (for even indices) */}
-//               {isTopPosition && (
-//                 <div className="absolute left-1/2 transform -translate-x-1/2" style={{ top: '-10px' }}>
-//                   <div 
-//                     className={`bg-white border-2 border-gray-200 rounded-lg shadow-md hover:shadow-lg transition-all duration-300 text-center cursor-pointer hover:scale-110 ${
-//                       expandedSquare === item ? 'px-7 py-5 scale-110 z-20' : 'px-5 py-3'
-//                     } min-w-max whitespace-nowrap`}
-//                     onClick={() => handleSquareClick(item)}
-//                   >
-//                     <span className="text-gray-700 text-sm sm:text-base font-medium block">
-//                       {timelineTexts[index] || `Step ${item}`}
-//                     </span>
-//                     {expandedSquare === item && (
-//                       <button
-//                         className="mt-2 px-3 py-1 text-white text-xs rounded transition-colors"
-//                         style={{
-//                           backgroundColor: clickedItems.has(item) ? '#c7b8e6' : '#a2e3cc'
-//                         }}
-//                         onMouseEnter={(e) => {
-//                           e.currentTarget.style.backgroundColor = clickedItems.has(item) ? '#b8a6d9' : '#8dd4b8';
-//                         }}
-//                         onMouseLeave={(e) => {
-//                           e.currentTarget.style.backgroundColor = clickedItems.has(item) ? '#c7b8e6' : '#a2e3cc';
-//                         }}
-//                         onClick={(e) => {
-//                           e.stopPropagation();
-//                           handleMarkAsDone(item);
-//                         }}
-//                       >
-//                         {clickedItems.has(item) ? 'Move back to in progress?' : 'Mark as done?'}
-//                       </button>
-//                     )}
-//                   </div>
-//                 </div>
-//               )}
+          return (
+            <div key={milestone.id} className="relative flex-shrink-0 mb-8" style={{ width: '120px', height: '270px', minWidth: '120px' }}>
+              {/* Square on top (for even indices) */}
+              {isTopPosition && (
+                <div className="absolute left-1/2 transform -translate-x-1/2" style={{ top: '-10px' }}>
+                  <div 
+                    className={`border-2 border-gray-200 rounded-lg shadow-md hover:shadow-lg transition-all duration-300 text-center cursor-pointer hover:scale-110 ${
+                      milestone.isCompleted ? 'bg-sky-300' : 'bg-neutral-300'
+                    } ${
+                      expandedSquare === milestone.id ? 'px-7 py-5 scale-110 z-20' : 'px-5 py-3'
+                    } min-w-max whitespace-nowrap`}
+                    onClick={() => handleSquareClick(milestone.id)}
+                  >
+                    <span className="text-gray-700 text-sm sm:text-base font-medium block">
+                      {milestone.title}
+                    </span>
+                    {milestone.isCompleted && milestone.completionDate && (
+                      <span className="text-xs mt-1 block" style={{ 
+                        color: 'black', 
+                        opacity: 0.8 
+                      }}>
+                        {formatDate(milestone.completionDate)}
+                      </span>
+                    )}
+                    {expandedSquare === milestone.id && (
+                      <button
+                        className="mt-2 px-3 py-1 text-white text-xs rounded transition-colors"
+                        style={{
+                          backgroundColor: milestone.isCompleted ? '#c7b8e6' : '#a2e3cc'
+                        }}
+                        onMouseEnter={(e) => {
+                          e.currentTarget.style.backgroundColor = milestone.isCompleted ? '#b8a6d9' : '#8dd4b8';
+                        }}
+                        onMouseLeave={(e) => {
+                          e.currentTarget.style.backgroundColor = milestone.isCompleted ? '#c7b8e6' : '#a2e3cc';
+                        }}
+                        onClick={(e) => {
+                          e.stopPropagation();
+                          handleMarkAsDone(milestone.id);
+                        }}
+                      >
+                        {milestone.isCompleted ? 'Move back to in progress?' : 'Mark as done?'}
+                      </button>
+                    )}
+                  </div>
+                </div>
+              )}
               
-//               {/* Timeline Circle - centered vertically and horizontally */}
-//               <div 
-//                 className="absolute left-1/2 transform -translate-x-1/2 w-14 h-14 rounded-full border-4 border-white shadow-lg flex items-center justify-center transition-all duration-300 z-10"
-//                 style={{
-//                   top: '50%',
-//                   marginTop: '-28px', // Half of circle height to center it
-//                   backgroundColor: clickedItems.has(item) ? '#a2e3cc' : '#fdfbf1'
-//                 }}
-//               >
-//                 <span className="text-gray-800 font-semibold text-base">
-//                   {item}
-//                 </span>
-//               </div>
+              {/* Timeline Circle - centered vertically and horizontally */}
+              <div 
+                className="absolute left-1/2 transform -translate-x-1/2 w-14 h-14 rounded-full border-4 border-white shadow-lg flex items-center justify-center transition-all duration-300 z-10"
+                style={{
+                  top: '50%',
+                  marginTop: '-28px', // Half of circle height to center it
+                  backgroundColor: milestone.isCompleted ? '#a2e3cc' : '#e5e7eb'
+                }}
+              >
+                <span className="text-gray-800 font-semibold text-base">
+                  {milestone.index}
+                </span>
+              </div>
 
-//               {/* Small arrow pointing from circle to square */}
-//               <div className="absolute left-1/2 transform -translate-x-1/2" style={{
-//                 top: isTopPosition ? 'calc(50% - 50px)' : 'calc(50% + 50px)',
-//                 opacity: 0.4
-//               }}>
-//                 <svg 
-//                   className="w-3 h-3" 
-//                   viewBox="0 0 12 12" 
-//                   fill="none"
-//                   style={{
-//                     transform: isTopPosition ? 'rotate(180deg)' : 'rotate(0deg)'
-//                   }}
-//                 >
-//                   <path 
-//                     d="M6 2L6 10M6 10L3 7M6 10L9 7" 
-//                     stroke="#6b7280" 
-//                     strokeWidth="1.5" 
-//                     strokeLinecap="round" 
-//                     strokeLinejoin="round"
-//                   />
-//                 </svg>
-//               </div>
+              {/* Small arrow pointing from circle to square */}
+              <div className="absolute left-1/2 transform -translate-x-1/2" style={{
+                top: isTopPosition ? 'calc(50% - 50px)' : 'calc(50% + 50px)',
+                opacity: 0.4
+              }}>
+                <svg 
+                  className="w-6 h-6" 
+                  viewBox="0 0 24 24" 
+                  fill="none"
+                  style={{
+                    transform: isTopPosition ? 'rotate(180deg)' : 'rotate(0deg)'
+                  }}
+                >
+                  <path 
+                    d="M12 4L12 20M12 20L6 14M12 20L18 14" 
+                    stroke="#6b7280" 
+                    strokeWidth="2" 
+                    strokeLinecap="round" 
+                    strokeLinejoin="round"
+                  />
+                </svg>
+                
+              </div>
 
-//               {/* Square on bottom (for odd indices) */}
-//               {!isTopPosition && (
-//                 <div className="absolute left-1/2 transform -translate-x-1/2" style={{ bottom: '-10px' }}>
-//                   <div 
-//                     className={`bg-white border-2 border-gray-200 rounded-lg shadow-md hover:shadow-lg transition-all duration-300 text-center cursor-pointer hover:scale-110 ${
-//                       expandedSquare === item ? 'px-7 py-5 scale-110 z-20' : 'px-5 py-3'
-//                     } min-w-max whitespace-nowrap`}
-//                     onClick={() => handleSquareClick(item)}
-//                   >
-//                     <span className="text-gray-700 text-sm sm:text-base font-medium block">
-//                       {timelineTexts[index] || `Step ${item}`}
-//                     </span>
-//                     {expandedSquare === item && (
-//                       <button
-//                         className="mt-2 px-3 py-1 text-white text-xs rounded transition-colors"
-//                         style={{
-//                           backgroundColor: clickedItems.has(item) ? '#c7b8e6' : '#a2e3cc'
-//                         }}
-//                         onMouseEnter={(e) => {
-//                           e.currentTarget.style.backgroundColor = clickedItems.has(item) ? '#b8a6d9' : '#8dd4b8';
-//                         }}
-//                         onMouseLeave={(e) => {
-//                           e.currentTarget.style.backgroundColor = clickedItems.has(item) ? '#c7b8e6' : '#a2e3cc';
-//                         }}
-//                         onClick={(e) => {
-//                           e.stopPropagation();
-//                           handleMarkAsDone(item);
-//                         }}
-//                       >
-//                         {clickedItems.has(item) ? 'Move back to in progress?' : 'Mark as done?'}
-//                       </button>
-//                     )}
-//                   </div>
-//                 </div>
-//               )}
+              {/* Square on bottom (for odd indices) */}
+              {!isTopPosition && (
+                <div className="absolute left-1/2 transform -translate-x-1/2" style={{ bottom: '-10px' }}>
+                  <div 
+                    className={`border-2 border-gray-200 rounded-lg shadow-md hover:shadow-lg transition-all duration-300 text-center cursor-pointer hover:scale-110 ${
+                      milestone.isCompleted ? 'bg-sky-300' : 'bg-neutral-300'
+                    } ${
+                      expandedSquare === milestone.id ? 'px-7 py-5 scale-110 z-20' : 'px-5 py-3'
+                    } min-w-max whitespace-nowrap`}
+                    onClick={() => handleSquareClick(milestone.id)}
+                  >
+                    <span className="text-gray-700 text-sm sm:text-base font-medium block">
+                      {milestone.title}
+                    </span>
+                    {milestone.isCompleted && milestone.completionDate && (
+                      <span className="text-xs mt-1 block" style={{ 
+                        color: 'black', 
+                        opacity: 0.8 
+                      }}>
+                        {formatDate(milestone.completionDate)}
+                      </span>
+                    )}
+                    {expandedSquare === milestone.id && (
+                      <button
+                        className="mt-2 px-3 py-1 text-white text-xs rounded transition-colors"
+                        style={{
+                          backgroundColor: milestone.isCompleted ? '#c7b8e6' : '#a2e3cc'
+                        }}
+                        onMouseEnter={(e) => {
+                          e.currentTarget.style.backgroundColor = milestone.isCompleted ? '#b8a6d9' : '#8dd4b8';
+                        }}
+                        onMouseLeave={(e) => {
+                          e.currentTarget.style.backgroundColor = milestone.isCompleted ? '#c7b8e6' : '#a2e3cc';
+                        }}
+                        onClick={(e) => {
+                          e.stopPropagation();
+                          handleMarkAsDone(milestone.id);
+                        }}
+                      >
+                        {milestone.isCompleted ? 'Move back to in progress?' : 'Mark as done?'}
+                      </button>
+                    )}
+                  </div>
+                  
+                </div>
+              )}
 
-//               {/* Arrow (except for last item) */}
-//               {index < items.length - 1 && (
-//                 <div className="absolute left-full top-1/2 transform -translate-y-1/2 flex items-center justify-center" style={{ width: '20px' }}>
-//                   <svg 
-//                     className="w-4 h-3 sm:w-6 sm:h-4" 
-//                     viewBox="0 0 24 16" 
-//                     fill="none"
-//                   >
-//                     <path 
-//                       d="M2 8h16m0 0l-4-4m4 4l-4 4" 
-//                       stroke="#6b7280" 
-//                       strokeWidth="2" 
-//                       strokeLinecap="round" 
-//                       strokeLinejoin="round"
-//                     />
-//                   </svg>
-//                 </div>
-//               )}
-              
-//               {/* Final Arrow (after last circle) */}
-//               {index === items.length - 1 && (
-//                 <div className="absolute left-full top-1/2 transform -translate-y-1/2 flex items-center justify-center ml-2">
-//                   <svg 
-//                     className="w-4 h-3 sm:w-6 sm:h-4" 
-//                     viewBox="0 0 24 16" 
-//                     fill="none"
-//                   >
-//                     <path 
-//                       d="M2 8h16m0 0l-4-4m4 4l-4 4" 
-//                       stroke="#c7b8e6" 
-//                       strokeWidth="2" 
-//                       strokeLinecap="round" 
-//                       strokeLinejoin="round"
-//                     />
-//                   </svg>
-//                 </div>
-//               )}
-//             </div>
-//           );
-//         })}
-//       </div>
-//     </div>
-//   );
-// }
+              {/* Arrow (except for last item) */}
+              {index < milestoneItems.length - 0 && (
+                <div className="absolute left-full top-1/2 transform -translate-y-1/2 flex items-center justify-center" style={{ width: '30px' }}>
+                  <svg 
+                    className="w-4 h-3 sm:w-6 sm:h-4" 
+                    viewBox="0 0 24 16" 
+                    fill="none"
+                  >
+                    <path 
+                      d="M2 8h16m0 0l-4-4m4 4l-4 4" 
+                      stroke="#6b7280" 
+                      strokeWidth="2" 
+                      strokeLinecap="round" 
+                      strokeLinejoin="round"
+                    />
+                  </svg>
+                </div>
+              )}
+            </div>
+          );
+        })}
+      </div>
+      )}
+    </div>
+  );
+}
